pyproject.toml
[tool.poetry]
name = "gdh2"
version = "0.1.0"
description = "Guardião das Horas 2.0 — controle de escala, banco de horas e ajuda de custo."
authors = ["Alex Almeida <you@example.com>"]
readme = "README.md"
packages = [{ include = "gdh2", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg = {version="^3.2.0", extras=["binary"]}
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.0"
black = "^24.8.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
